refactor(tests): move detokenizer tests to integration tests
Moved detokenizer tests from src/lex/formats/detokenizer/detokenizer.rs
to tests/detokenizer/main.rs following Rust convention.

Changes:
- Created tests/detokenizer/main.rs with integration tests
- Tests now use public API exclusively (detokenize, tokenize, lex)
- Migrated all snapshot files to tests/detokenizer/snapshots/
- Updated snapshot naming to match integration test convention
- Removed unit tests from src/ (kept only public re-exports)
- All 343 tests pass

This follows the Rust convention where integration tests in tests/
directory test the public API as external users would, while unit
tests in src/ test internal implementation details.
